Save non-default currency to database and display it on fetch for Booking

+++ src/main/java/at/ac/fhcampuswien/car_rental/dao/car/CarEntity.java
@@ -10,6 +10,7 @@ import lombok.experimental.FieldDefaults;
 import org.hibernate.validator.constraints.Length;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.util.Set;
 
 @Entity
@@ -31,7 +32,7 @@ public class CarEntity extends BaseEntity {
     @Column(name = "NUM_CONSTRUCTION_YEAR")
     Integer constructionYear;
     @Column(name = "BI_PRICE")
-    Float price;
+    BigDecimal price;
     @Column(name = "TXT_CURRENCY")
     String currency;
     @OneToMany(mappedBy="car")
